requirements.txt
django ~=3.0.0
djangorestframework
djangorestframework-jwt
django-cors-headers
celery[redis]
django_celery_beat
django_celery_results
requests